+++ libs/ngx-material-intl-tel-input-lib/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/ngx-material-intl-tel-input-lib/ngx-material-intl-tel-input-lib.component';
 export * from './lib/types/text-labels.type';
+export * from './lib/enums/country-iso.enum';