+++ libs/ngx-material-intl-tel-input-lib/src/lib/services/geo-ip/geo-ip.service.spec.ts
@@ -1,12 +1,15 @@
 import { TestBed } from '@angular/core/testing';
-
 import { GeoIpService } from './geo-ip.service';
+import { HttpClientModule } from '@angular/common/http';
 
 describe('GeoIpService', () => {
   let service: GeoIpService;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
+    TestBed.configureTestingModule({
+      imports: [HttpClientModule],
+      providers: [GeoIpService]
+    });
     service = TestBed.inject(GeoIpService);
   });