+++ libs/ngx-material-intl-tel-input-lib/src/lib/types/country.model.ts
@@ -1,10 +1,12 @@
+import { CountryISO } from '../enums/country-iso.enum';
+
 export type Country = {
-	name: string;
-	iso2: string;
-	dialCode: string;
-	priority: number;
-	areaCodes?: string[];
-	htmlId: string;
-	flagClass: string;
-	placeHolder: string;
-}
+  name: string;
+  iso2: CountryISO | string;
+  dialCode: string;
+  priority: number;
+  areaCodes?: string[];
+  htmlId: string;
+  flagClass: string;
+  placeHolder: string;
+};